requirements.txt
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
urllib3==2.0.6
